# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Financial Ledger API",
        "description": "Ledger-based financial reporting API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default"
        }
    ],
    "paths": {
        "/api/financial-reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Aggregate financial ledger report for a date range",
                "operationId": "d0ec184b99d7808a74d2be53b0d4c2aa",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/financial-reports/chart": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Time-series chart data for income/expense in a date range",
                "operationId": "f9505796943a48e568e85ad340a63f89",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "granularity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "day",
                                "month"
                            ],
                            "example": "day"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show a single order",
                "operationId": "f090f23952a43521af6f1d133826bc37",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a pending order",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "total_amount"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "total_amount": {
                                        "type": "integer",
                                        "example": 100000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/orders/{order}/pay": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a mock payment URL for an order",
                "operationId": "d4190031f37fe1e0cb1d620ed2236017",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "409": {
                        "description": "Order not payable"
                    }
                }
            }
        },
        "/api/orders/callback": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Mock bank callback to finalize payment",
                "operationId": "2ba2a4faca53e6585aab5a9358f8a85e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "status"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "success",
                                            "failed"
                                        ],
                                        "example": "success"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Processed"
                    },
                    "409": {
                        "description": "Order already completed/cancelled"
                    },
                    "422": {
                        "description": "Payment failed or validation error"
                    }
                }
            }
        },
        "/api/transactions/{transaction}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Show a single transaction",
                "operationId": "05e2714a413c88baa08631aea0c3bb87",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "List transactions (paginated)",
                "operationId": "6581469356772e7a74a0d79897544f1a",
                "parameters": [
                    {
                        "name": "wallet_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "platform_commission"
                        }
                    },
                    {
                        "name": "flow",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "in",
                                "out"
                            ]
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/wallets/withdraw": {
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Withdraw from a wallet (creates an OUT transaction and decrements balance)",
                "operationId": "c578418ba6480b5273a41ba51d321a3e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "wallet_id",
                                    "amount"
                                ],
                                "properties": {
                                    "wallet_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 50000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "User withdrawal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation/insufficient balance"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Order creation and payment callbacks"
        },
        {
            "name": "Reports",
            "description": "Financial reporting endpoints"
        },
        {
            "name": "Transactions",
            "description": "Ledger transactions"
        },
        {
            "name": "Wallets",
            "description": "Wallet operations"
        }
    ]
}